back/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19
pydantic[email]==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
